+++ src/react-native/components/context/EditorContext.tsx
@@ -1,4 +1,9 @@
-import { Bridge, BuiltinBridgeKey, ReactNativeBridgeToken } from '../../utils';
+import {
+  Bridge,
+  BridgeRegistryKey,
+  BuiltinBridgeKey,
+  WebViewBridgeSDK,
+} from '../../utils';
 import React, { FC, PropsWithChildren } from 'react';
 import type { ResolverList } from '../../utils';
 import type { WebView } from 'react-native-webview';
@@ -10,13 +15,12 @@ export interface IEditorContextProps {
 }
 export interface IEditorContextValue extends IEditorContextProps {
   bridges: Map<string, Bridge<any, any>>;
-  getBridge: (key: string) => Bridge<any, any> | undefined;
+  getBridge: <S extends ResolverList, T extends ResolverList>(
+    key: string
+  ) => Bridge<S, T> | undefined;
   injectJavaScript: (script: string) => Promise<void>;
 }
-interface IBridgeRegisterProps {
-  registerKey: string;
-  resolverList?: ResolverList;
-}
+
 const initialValue: IEditorContextValue = {
   webViewRef: React.createRef(),
   isEditorReady: false,
@@ -26,56 +30,35 @@ const initialValue: IEditorContextValue = {
   injectJavaScript: () => Promise.resolve(),
 };
 export const EditorContext = React.createContext(initialValue);
-const wrapper = () => {
-  const bridges = new Map<string, Bridge<any, any>>();
-  const EditorContextProvider: FC<PropsWithChildren<IEditorContextProps>> = (
-    props
-  ) => {
-    const injectJavaScript = React.useCallback(
-      async (script: string) => {
-        const bridge__builtin = bridges.get(BuiltinBridgeKey);
-        if (bridge__builtin) {
-          const { promise, id } = bridge__builtin.createPromise();
-          props.webViewRef.current?.injectJavaScript(
-            `${script};${ReactNativeBridgeToken}.resolvePromise('${id}')`
-          );
-          await promise;
-        }
-      },
-      [props.webViewRef]
-    );
-    const value: IEditorContextValue = React.useMemo(() => {
-      return {
-        bridges,
-        getBridge: (key: string) => bridges.get(key),
-        injectJavaScript,
-        ...props,
-      };
-    }, [injectJavaScript, props]);
+export const EditorContextProvider: FC<
+  PropsWithChildren<IEditorContextProps>
+> = (props) => {
+  const bridges = React.useRef(new Map<string, Bridge<any, any>>());
+  const injectJavaScript = React.useCallback(
+    async (script: string) => {
+      const bridge__builtin = bridges.current.get(BuiltinBridgeKey);
+      if (bridge__builtin) {
+        const { promise, id } = bridge__builtin.createPromise();
+        props.webViewRef.current?.injectJavaScript(
+          `${script};${WebViewBridgeSDK}.${BridgeRegistryKey}.get('${BuiltinBridgeKey}').resolvePromise('${id}')`
+        );
+        await promise;
+      }
+    },
+    [props.webViewRef]
+  );
+  const value: IEditorContextValue = React.useMemo(() => {
+    return {
+      bridges: bridges.current,
+      getBridge: (key: string) => bridges.current.get(key),
+      injectJavaScript,
+      ...props,
+    };
+  }, [injectJavaScript, props]);
 
-    return (
-      <EditorContext.Provider value={value}>
-        {props.children}
-      </EditorContext.Provider>
-    );
-  };
-
-  const BridgeRegister: FC<IBridgeRegisterProps> = (props) => {
-    const { registerKey, resolverList } = props;
-    const registered = React.useRef(false);
-    if (!registered.current) {
-      bridges.set(registerKey, new Bridge<any, any>(resolverList));
-      registered.current = true;
-    }
-    return <></>;
-  };
-
-  return {
-    EditorContextProvider,
-    BridgeRegister,
-  };
+  return (
+    <EditorContext.Provider value={value}>
+      {props.children}
+    </EditorContext.Provider>
+  );
 };
-
-const components = wrapper();
-export const EditorContextProvider = components.EditorContextProvider;
-export const BridgeRegister = components.BridgeRegister;