+++ src/react-native/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useEditorContext';
 export * from './useBridge';
 export * from './useFocus';
 export * from './useFormat';
+export * from './useBridgeRegister';