+++ src/react-native/hooks/useBridge.ts
@@ -1,19 +1,18 @@
 import React from 'react';
 import type {
-  Bridge,
   QuillResolversBuiltin,
   ResolverList,
   RNResolversBuiltin,
   WebViewResolversBuiltin,
 } from '../utils';
-import { BuiltinBridgeKey } from '../utils';
+import { Bridge, BuiltinBridgeKey } from '../utils';
 import { useEditorContext } from './useEditorContext';
 export const useBridge = <SRC extends ResolverList, TGT extends ResolverList>(
   registerKey: string
 ) => {
   const { getBridge } = useEditorContext();
   return React.useMemo(() => {
-    return getBridge(registerKey) as Bridge<SRC, TGT>;
+    return getBridge<SRC, TGT>(registerKey) as Bridge<SRC, TGT>;
   }, [getBridge, registerKey]);
 };